e2e/requirements.txt
pytest
pytest-asyncio
playwright
boto3
python-dotenv
requests
